unsgine long int

+++ bit_manipulation/5-flip_bits.c
@@ -9,7 +9,7 @@
 unsigned int flip_bits(unsigned long int n, unsigned long int m)
 {
 	int count = 0;
-	unsigned int s = n ^ m;
+	unsigned long int s = n ^ m;
 
 	while (s > 0)
 	{
